feat(#752): remove the old implementation of MaxLocals
